refactor: separate app config and server entry point
